requirements.txt
Django==4.2.7
mysqlclient==2.2.0
Pillow==10.0.1
python-decouple==3.8
whitenoise==6.6.0
